Fix season manager crashes and flag rendering
- Guard new-season creation against a null actual calendar in
  SimulationSaveManagerWindow::configNextSeason and clear all calendar UI
- Reorder calendar editor add/duplicate to insert data before the row
  insert signal so rows appear immediately
- Null-guard ClassificationResultsTableModel headerData/data
- Scale start list and KO group result flag icons to 24x14
- Guard competition config against empty rounds
- Make Windows qmake builds portable (drop DPP, Qt Charts namespaces)
- Fix KO group min() width-cast warning

+++ source/competitions/KOSystem/KOGroup.cpp
@@ -127,7 +127,7 @@ QVector<KOGroup> KOGroup::constructKOGroups(RoundInfo *roundInfo, QVector<Jumper
         for(int basketIndex = 0; basketIndex < groupSize; ++basketIndex)
         {
             const int from = basketIndex * groupsCount;
-            const int to = std::min(from + groupsCount, jumpers->count());
+            const int to = std::min(from + groupsCount, static_cast<int>(jumpers->count()));
             QVector<Jumper *> basket;
             for(int index = from; index < to; ++index)
                 basket.push_back(jumpers->at(index));
